Moved config package out of internal directory
We experienced some problems with running tests in docker image when
config package was under internal directory.
There is no strong reason for having it in internal directory. Moving it
to root directory is what we have in keep-core and to stay consistent
we're aligning this code with it.
