+++ pkg/client/config.go
@@ -3,7 +3,7 @@ package client
 import (
 	"time"
 
-	configtime "github.com/keep-network/keep-ecdsa/internal/config/time"
+	configtime "github.com/keep-network/keep-ecdsa/config/time"
 )
 
 const (
